+++ compiler-plugin/src/main/java/io/ballerina/stdlib/persist/compiler/Constants.java
@@ -31,6 +31,8 @@ public final class Constants {
     public static final String EMPTY_STRING = "";
     public static final String ARRAY = "[]";
     public static final String LS = System.lineSeparator();
+    public static final String SQL_RELATION_MAPPING_ANNOTATION_NAME = "sql:Relation";
+    public static final String ANNOTATION_REFS_FIELD = "refs";
 
     private Constants() {
     }